First of probably many passes for making parameter evaluation thread-safe. This commit started out as a pass to make ev_errno become a thread local storage variable via UT_ThreadSpecificValue. However, there are performance issues in using UT_ThreadSpecificValue in its call to UT_Thread::getMySequentialThreadIndex().
So in an attempt to ameliorate this cost, this commit adds back in use of the
"thread" parameter in the parameter evaluation functions. I removed all default
parameters to functions for this parameter, and made sure (as much as I could)
that it is propagated along.

The (new) semantics for the "thread" variable is to use the value you were
passed in if you have it. Else, use the newly added OP_Context::getThread()
function to obtain a value for it if you have an OP_Context. Failing that,
you can call UT_Thread::getMySequentialThreadIndex(), or the newly added
pseudonym for it, UTgetSTID().

At the same time, I've also made use of the "thread" parameter as an index to
any of the UT_ThreadSpecificValue variables along the way.

$UT/UT_Thread.h
$UT/UT_ThreadSpecificValue.h
    - Since the default value for "thread" used to be 0, I've bumped up
      sequential thread indices by 1 in UT_Thread so that is the new minimum
      sequential thread id. In UT_ThreadSpecificValue, we account for this by
      subtracting 1 to index into our array.
    - At the same time, I factored out some duplicate code in
      UT_ThreadSpecificValue.
    - In UT_Thread.h, added UTgetSTID(), which simply calls
      UT_Thread::getMySequentialThreadIndex()

$EXPR/EXPR.h
$EXPR/EX_Error.h
$EXPR/Makefile
    Factored out the error handling functions into a new EX_Error.h file so
    that I can more easily inline its usage without pulling everything else.

$OP/OP_Parameters.h
    - Added *T() versions of various functions that take a thread parameter.
      If you call the old version, then it calls the new version with
      UTgetSTID() as the thread parameter.
	- followChannelReferencesOneStepT()
	- followChannelReferencesT()
	- eval*T()
    - For performance, made checkExpressionError() always inline and at the
      same time, made evaluationError() private since no else was calling it.

Everthing else
    Added the thread parameter and pass it through. In some cases, I changed
    a raw PRM_Parm::getValue() call into the corresponding OP_Parameter's eval
    call.