Consolidated build components, whose inclusion was previosly controlled from multiple locations, into a single file in $SHM/Components.
The current set of components is: FBX, OpenCL and OpenVDB.

Each of those components is enabled depending on platform availability and suitability. Each can also be forcibly disabled. If a component is enabled,an appropriate make flag is set ({FBX,OPENCL,OPENVDB}_ENABLED) so that the correct behaviour can be defined in the make files. A define of the same name is added to the compile flags to control behaviour on the source level.