Adding better support for precompiled headers. First, enabled precompiled headers on Mac and Linux. The performance benefit on those platforms may not be as drastic as on Windows. But I'm not sure my testing in a VM is really comparable to a real machine. On all platforms, precompiled headers now apply only to $SHL. By default the same precompiled header is used for all libs, resulting in a marginal speedup. This is the same precompiled header support we already have on Windows. However if you have about 30GB of extra disk space, you can set "HOUDINI_BUILD_PER_LIB_PCH=1" to use the precompiled.h file in each lib directory. On my Windows box this let me do a complete rebuild of $SHL in 20 minutes (with FORCE_PARALLEL turned on as well) instead of over 70 minutes with the single precompiled header.
I tried a build with and without this define on both Windows and Linux with
no issues. I'm relying on the assumption that Mac builds also use gcc and
so will support the same command line options as we use on Linux for
precompiled headers. Apologies to Mac developers if this is not the case.
I also was unable to try a VC 2012 build, but I don't believe that we
officially support that version now anyway.

list_link_libs.py got a new "-p" option to create the precompiled header
for a particular library. Each library has a set of header files that show
up in the precompiled header of any library that depends on it. A few libs
also have a header inside that lib precompiled (SOP_Node.h in SOP, for
example). You can do "make precompiled_header" in $SHL or below to
generate the precompiled header for a given lib. If you want to change
a precompiled header, please do so through list_link_libs.py rather than
editing the precompiled.h directly.

Changes to FUSE, GT, SOP, PRM, SIM, and BV were renaming enums to avoid
conflicting with the same symbols used elsewhere in the code.

Change to ROP_FBXHeaderWrapper and FBX_HeaderWrapper were to undefine the
symbol PropertyNotify which in X.h is defined as a constant. But FBX uses
it as a function name.

UT_IPC.h was missing an include of winsock.h on Windows (it relied on that
file being included in the C file before this header).

Updated the Windows and linux build filters to say "Compiling precompiled.h"
whien building a precompiled header.