Remove the old network dot implementation, and replace it with a totally different one. In the new implementation, network dots are a subclass of OP_IndirectInput (with OP_SubnetIndirectInput, aka OP_InputIndirect being the other subclass). This provides a common code path for handling the multiple outputs that can emerge from an indirect input. But OP_Dot also has an OP_Input o something can be wired directly into it (only one thing though).
This means a network dot can exist in isolation, with just an input, or with
just an output. It has a name, but that name is only used for scripting, and
as a way to differentiate between different dots in an OPUI_InputEditor.

Because dots should be essentially "transparent" to nodes that connect into
them, the getNumOutputs and getOuputs functions on OP_Node have been removed,
since it is no longer practical to iterate through all the nodes connected to
the output of another node. Instead there is an OP_OutputIterator (and an
OP_OutputReversedIterator) class you can put into a range based for loop that
lets you iterate over all the output nodes (treating dots as transparent).
Then there are the nOutputItems() and getOutputItem() which return the things
actually connected to the node (be they nodes or dots).

Took advantage of the sweeps through here (especially in collapsing and
expanding subnets) to consolidate some code that handled postits, netboxes,
indirect inputs, nodes, and dots all as separate cases. It is now possible
to look up any OP_NetworkBoxItem by name/path. Everything with a name was
already sharing a single namespace, and now indirect inputs are identified
by the names "1", "2", "3", etc. Which avoids the namespace issue because
those are not legal names for any other item type.

Made a pretty good first pass over the NNE python code to handle the new dots
in reasonable ways. There are still some holdovers that don't work like the
RMB context menu on a dot. Dots have their own equivalent of the flyout ring
which just has two dots (input and output) to let you make connections
starting from a dot. Once a connection has been started, a dot ia always a
valid destination, becaues it can always be an input or an output.

The ONE should mostly ignore the new dots. If it doesn't, and maybe crashes
in some cases, I don't really care.

Any dots you may have in old hip files will not be ported over. They are gone.
Apologies for that.