from dev trunk: Fix some of the FBX capture weight import/export issues as found in bug 97968: - In FBX_DerivedActions.C, guard against dividing by 0 - In ROP_FBXMainVisitor.C, set skin_deform_node to nullptr as soon as we've   output some geometry with it to avoid outputting it multiple times.
The actual problem here is that when we output the capture weights, it's
assumed that the point indexing in Houdini matches that of the FBX shape that
is being output. However, that's not always the case as evidenced by
ROP_FBXMainVisitor::outputSOPNodeWithoutVC()'s GEOPRIMPOLY case where if there
are any open polygons (aka polylines), then it outputs them separately. So
because we run into this with bug 97968, we end up having the wrong point
indexing for capture weights.