from dev trunk so that I can backport some later improvements without running into conflicts:
Add some rudimentary support for exporting Switcher cameras to FBX, where only
transform can be animated. A proper way to do this is to actually create an
FBX camera switcher node but this lays in some infrastructure for it.

To do this, I moved ROP_FBXnetworkTypesToIgnore functionality to the actual
only place where it's ALWAYS used: ROP_FBXBaseVisitor::isNetworkVisitable().
As a result removed:
- ROP_FBXnetworkTypesToIgnore
- ROP_FBXBaseVisitor::addNonVisitableNetworkType()
- ROP_FBXBaseVisitor::addNonVisitableNetworkTypes()
- ROP_FBXBaseVisitor::myNetworkTypesNotToVisit

Fix r354276's ROPfbxCastToCamera() so that it doesn't erroneously flag lights
as cameras.  This was happening because OBJ_Light is a descendant subclass of
OBJ_Camera.