from dev trunk:
Add ability for the Filmbox FBX ROP to override the exported axis system
without converting from the native Houdini axis system by adding a parm toggle
to convert that defaults to off. The Axis System menu now has an additional
menu item of "Current" which means to do export to whatever orientation the
main preferences are using.

To support this, moved the storing of GUI_ViewParameter::myOrientationMode down
to OP_Director so that we can access this from ROP_FBX without needing to link
against any of the UI libraries.