from dev trunk:
Add Convert Units parameter to FBX Filmbox ROP/SOP. When enabled, it will
export by first converting to centimeters from the Houdini system units.
It defaults to off so that it matches the old behaviour and the FBX scene
importer (so that we can round trip without when using scene FBX
importer/exporter).