from dev trunk:
Add in the same hacky heuristic as the non-path specific exporter to export
nodes as LODGroup nodes instead of Null. Factored out a new
ROP_FBXUtil::isLODGroupNullNodeName() method to do this detection.
This is far from complete though since we currently have no way to specify the
custom properties needed for them.